dev: Keep srv directory to bring up container
The PHP container demand the srv directory
because it serves the website.
This directory is not there out of the box
which makes it impossible for the PHP container to start
without some hands on.
We do not want to trouble the developers
so we add the directory in the VCS.