dev: Refactor FixtureCommand for more OOP
The current implementation of FixtureCommand is a bit quick and dirty.
From time to time we try to decouple as much logic as possible from this.
This time a function parameter has been swapped to a class param
so that we can delegate/swap the loops in the __invoke method some time.