ci: Less runner for faster results
CI runs with almost any possible combination.
This consumes a lot of time
and feels to much on the PHP side.
Therefor we only test the edges with all
PHP versions
and mix PHP versions up in between.
As PHP has the best backward compatibility / semver
among all candidates we are sure that this coverage is enough.