wp: wp_is_site_initialized compatibility
The function wp_is_site_initialized has been introduced in WP 5.1.0
and is used for test-assertions.
As we are trying to support lower WP versions (>=4.6) this test
always fails because the function wp_is_site_initialized is missing.
We copied that function into our base test case
and use it as a fallback,
if "wp_is_site_initialized" is not present.