hotfix: Old WP return stdClass over WP_Post when deleting post
When deleting a post the wp_delete_post function returns the deleted
post.
In some older WP versions it did not return this as a WP_Post object
but as a stdClass,
which made our check, if the post has been deleted, fail.
So instead of the type whe now check for FALSE
and the ID of the deleted post.