cli: Checking for compatibility before running wp-fixture
WordPress and PHP have some versions that they are incompatible with
(e.g. PHP 8.0 is only supported by WP 5.1 and higher).
This problem can not be covered by our composer.json version constraints
when using the "conflict" config.
So we added a compatibility check when running the wp-fixture command
from CLI / in wp-cli.