bc: Deprecated usage of fixture-names in repositories
The repositories got the fixture-name to produce error-messages.
Passing this information has become obsolete
because only one fixture is passed at a time to the repository,
so an exception is clearly related to the passed entity / fixture.
The obsolete parameter has been marked deprecated
and may be removed in future commits.