dev: Split TestCases in more simple versions
The current TestCase-Base is autoloading all the fixtures
that are defined in the doc-comment.
Sometimes there are no fixtures
or we define some but don't want them in the database.
So we needed to distinguish if the test should autoload all fixtures
or if we are on a more basic level with no need for that.

* TestCase.php: Deprecated, the more specific cases shall be used
* Added some traits to separate logic even more
  and not mess up the actual BasicTestCase with too much
  (possibly temporary) stuff